+++ dotNet/GoT/GoT.Server/Messages/GameDto.cs
@@ -7,11 +7,11 @@ namespace GoT.Server.Messages
 {
     public class GameDto
     {
-        public Int64 GameId { get; set; }
+        public Int32 GameId { get; set; }
         public DateTime Date { get; set; }
         public string Mode { get; set; }
         public string Status { get; set; }
         public DateTime CreatedDate { get; set; }
-        public Int64 CreatedByPlayerId { get; set; }
+        public Int32 CreatedByPlayerId { get; set; }
     }
 }
