+++ dotNet/GoT/GoT.Server/Messages/HouseDto.cs
@@ -7,12 +7,12 @@ namespace GoT.Server.Messages
 {
     public class HouseDto
     {
-        public Int64 HouseId { get; set; }
+        public Int32 HouseId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public byte[] Sigil { get; set; }
-        
-        public Int64 CapitalRegionId { get; set; }
+
+        public Int32 CapitalRegionId { get; set; }
         public List<CharacterDto> Characters { get; set; }
         public List<HouseCardDto> HouseCards { get; set; }
     }
