+++ dotNet/GoT/GoT.Server/Messages/PortDto.cs
@@ -7,7 +7,7 @@ namespace GoT.Server.Messages
 {
     public class PortDto
     {
-        public Int64 PortId { get; set; }
+        public Int32 PortId { get; set; }
         public string Name { get; set; }
     }
 }
